Minimum working shell completed.
Shell can now take commands from the user, search for them and then execute them. User can add their own paths to a file. Executables are being searched from this file. If not found, the shell simply moves on to the next command after showing some message to the user.